refactor: centralize document listing queries in Document model
- Create DocumentModel.getDocumentVersionsForListing() to eliminate code duplication
- Update both /docs and /users/[userId]/documents pages to use the centralized method
- Fix missing DocumentsClient import error in user documents page
- Improve maintainability by having single source of truth for complex Prisma queries
