+++ apps/web/src/models/Document.ts
@@ -891,6 +891,134 @@ export class DocumentModel {
     return dbDocs.map((dbDoc) => DocumentModel.formatDocumentFromDB(dbDoc));
   }
 
+  /**
+   * Get document versions for listing pages (optimized for DocumentsResults component)
+   * This method returns serialized data ready for client components
+   */
+  static async getDocumentVersionsForListing(options?: {
+    userId?: string;
+    searchQuery?: string;
+    limit?: number;
+  }) {
+    const { userId, searchQuery, limit = 50 } = options || {};
+
+    // Build where clause
+    const whereClause: any = {};
+    
+    if (userId) {
+      whereClause.document = { submittedById: userId };
+    }
+    
+    if (searchQuery?.trim() && searchQuery.trim().length >= 2) {
+      whereClause.searchableText = {
+        contains: searchQuery.trim().toLowerCase(),
+        mode: "insensitive",
+      };
+    }
+
+    // Execute query with consistent structure
+    const rawDocuments = await prisma.documentVersion.findMany({
+      where: whereClause,
+      include: {
+        document: {
+          include: {
+            evaluations: {
+              include: {
+                agent: {
+                  include: {
+                    versions: {
+                      orderBy: { version: "desc" },
+                      take: 1,
+                    },
+                  },
+                },
+                versions: {
+                  orderBy: { version: "desc" },
+                  take: 1,
+                  include: {
+                    comments: {
+                      include: {
+                        highlight: true,
+                      },
+                    },
+                    job: true,
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+      orderBy: { createdAt: "desc" },
+      take: limit,
+    });
+
+    // Serialize to plain objects for client component
+    return rawDocuments.map((doc) => ({
+      id: doc.id,
+      title: doc.title,
+      authors: doc.authors,
+      content: doc.content,
+      urls: doc.urls,
+      platforms: doc.platforms,
+      intendedAgents: doc.intendedAgents,
+      importUrl: doc.importUrl,
+      createdAt: doc.createdAt.toISOString(),
+      updatedAt: doc.updatedAt.toISOString(),
+      document: {
+        id: doc.document.id,
+        publishedDate: doc.document.publishedDate.toISOString(),
+        createdAt: doc.document.createdAt.toISOString(),
+        updatedAt: doc.document.updatedAt.toISOString(),
+        submittedById: doc.document.submittedById,
+        evaluations: doc.document.evaluations.map((evaluation) => ({
+          id: evaluation.id,
+          agentId: evaluation.agentId,
+          createdAt: evaluation.createdAt.toISOString(),
+          agent: {
+            id: evaluation.agent.id,
+            versions: evaluation.agent.versions.map((version) => ({
+              id: version.id,
+              name: version.name,
+              description: version.description,
+              providesGrades: version.providesGrades,
+            })),
+          },
+          versions: evaluation.versions.map((version) => ({
+            id: version.id,
+            grade: version.grade,
+            comments: version.comments.map((comment) => ({
+              id: comment.id,
+              description: comment.description,
+              importance: comment.importance,
+              grade: comment.grade,
+              highlight: {
+                id: comment.highlight.id,
+                startOffset: comment.highlight.startOffset,
+                endOffset: comment.highlight.endOffset,
+                prefix: comment.highlight.prefix,
+                quotedText: comment.highlight.quotedText,
+                isValid: comment.highlight.isValid,
+                error: comment.highlight.error,
+              },
+            })),
+            job: version.job
+              ? {
+                  priceInDollars: version.job.priceInDollars
+                    ? Number(version.job.priceInDollars)
+                    : null,
+                  llmThinking: version.job.llmThinking,
+                }
+              : null,
+            summary: version.summary,
+            analysis: version.analysis,
+            selfCritique: version.selfCritique,
+          })),
+        })),
+      },
+    }));
+  }
+
   static async create(data: {
     title: string;
     authors: string;
